Extract store from password provider
